# refactor(digest:sha2): optimizing sha2 implementation