inlining/unrolling message scheduling and compression function

* Intel Xeon E5-1650 v3 (Haswell, 64bit Linux):
  sha256: 190 MB/s => 190 MB/s
  sha512: 282 MB/s => 354 MB/s
* AMD A10-7850K (Kaveri, 64bit Linux):
  sha256: 120 MB/s => 151 MB/s
  sha512: 182 MB/s => 244 MB/s