@homu 2016-01-15
:pushpin: Commit 00abda2 has been approved by `hauleth`

<!-- @homu r=hauleth 00abda27646b24c3ab1c9be854b68c91b8821131 -->