@homu 2016-01-15
:zap: Test exempted - [status](https://travis-ci.org/libOctavo/octavo/builds/102665121)